added compression; lowered the precision from float64 to float32

+++ er3t/rtm/lrt/lrt_flx.py
@@ -439,10 +439,10 @@ class lrt_read_uvspec_flx:
         Nz = len(inits)
 
         self.dims      = {'X':'Wavelength', 'Y':'Altitude', 'Z':'Files'}
-        f_down_direct  = np.zeros((Nx, Ny, Nz), dtype=np.float64)
-        f_down_diffuse = np.zeros((Nx, Ny, Nz), dtype=np.float64)
-        f_down         = np.zeros((Nx, Ny, Nz), dtype=np.float64)
-        f_up           = np.zeros((Nx, Ny, Nz), dtype=np.float64)
+        f_down_direct  = np.zeros((Nx, Ny, Nz), dtype=np.float32)
+        f_down_diffuse = np.zeros((Nx, Ny, Nz), dtype=np.float32)
+        f_down         = np.zeros((Nx, Ny, Nz), dtype=np.float32)
+        f_up           = np.zeros((Nx, Ny, Nz), dtype=np.float32)
 
 
         for i, init in enumerate(inits):
